refactor: replace inject with hook useStores
