+++ schedule/src/App.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Provider } from 'mobx-react';
 
-import Settings from './mobx/settings';
+import Settings from './mobx/store/settings';
 import Test from './components/Test/Test';
 import './App.css';
