+++ schedule/src/components/Test/Test.tsx
@@ -1,19 +1,21 @@
 import React from 'react';
-import { observer, inject } from 'mobx-react';
+import { observer } from 'mobx-react';
 
-const tes = (prop: any) => {
-  prop.toggleHideOldEvents();
+import useStores from '../../mobx/context';
+import { ISettings } from '../../interfaces/settings/settings';
+
+const tes = (settings: ISettings) => {
+  console.log(settings.viewMode);
 };
 
-const Test = inject('settings')(
-  observer(({ settings }: any) => {
-    return (
-      <>
-        {/* eslint-disable-next-line jsx-a11y/click-events-have-key-events,jsx-a11y/no-static-element-interactions */}
-        <div onClick={() => tes(settings)}>test</div>
-      </>
-    );
-  }),
-);
+const Test = observer(() => {
+  const { settings } = useStores();
+  return (
+    <>
+      {/* eslint-disable-next-line jsx-a11y/click-events-have-key-events,jsx-a11y/no-static-element-interactions */}
+      <div onClick={() => tes(settings.settings)}>test</div>
+    </>
+  );
+});
 
 export default Test;
