+++ schedule/src/constants/settings.ts
@@ -1,6 +1,6 @@
 export const DEFAULT_TIMEZONE = 180;
 export const LOCAL_SETTINGS = 'scheduleSettingsTeam34';
-export const enum ScheduleView {
+export enum ScheduleView {
   Table,
   List,
   Calendar,
