+++ schedule/src/mobx/context.ts
@@ -1,0 +1,8 @@
+import React from 'react';
+import { MobXProviderContext } from 'mobx-react';
+
+const useStores = () => {
+  return React.useContext(MobXProviderContext);
+};
+
+export default useStores;
